requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cycler==0.10.0
Cython==0.29.21
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.4
django-appconf==1.0.4
django-bootstrap-modal-forms==2.0.1
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
gunicorn==20.0.4
install==1.3.4
jdcal==1.4.1
joblib==0.17.0
kiwisolver==1.3.1
MarkupPy==1.14
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
pep517==0.9.1
Pillow==8.0.1
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
threadpoolctl==2.1.0
toml==0.10.2
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
